pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = "learning"
authors = ["AhmeedDahy <ahmedss2235555@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langchain = "^0.3.12"
qdrant-client = "^1.12.1"
python-decouple = "^3.8"
openai = "^1.58.1"
langchain-community = "^0.3.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
